[staccmr-40] initial setup
