+++ search/lib/convert/index.js
@@ -1,4 +1,10 @@
 const boundingBox = require('./bounding-box');
 const collections = require('./collections');
+const granuleToItem = require('./granule-to-item');
 
-module.exports = Object.assign({}, boundingBox, collections);
+// module.exports = Object.assign({}, boundingBox, collections);
+module.exports = {
+  ...boundingBox,
+  ...collections,
+  ...granuleToItem
+};
